# public/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Book Store",
        "description": "Book Store is an APP built using NodeJS and MYSQL \n"
    },
    "host": "",
    "basePath": "/api",
    "schemes": [
        "http", "https"
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Returns access token on successful login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "request body parameters need to be added here",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/Login"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Sucess"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Authentication failure"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Returns list of all users",
                "parameters": [{
                    "in": "header",
                    "name": "Authorization",
                    "description": "Bearer followed by auth token recieved after login",
                    "required": true,
                    "type": "string",
                    "example": "Bearer eyJhbGciOiJIUzI1NiJ9.ZW1haWxAbm9lbWFpbC5jb20.fGbjTahMFHe9Ad330JH5HfkhNuP4FdXICuQHgKS9Xag"
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Sucess"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Authentication failure"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/getUserWithId": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Returns user with matching id ",
                "parameters": [{
                        "in": "header",
                        "name": "Authorization",
                        "description": "Bearer followed by auth token recieved after login",
                        "required": true,
                        "type": "string",
                        "example": "Bearer eyJhbGciOiJIUzI1NiJ9.ZW1haWxAbm9lbWFpbC5jb20.fGbjTahMFHe9Ad330JH5HfkhNuP4FdXICuQHgKS9Xag"

                    },
                    {
                        "in": "query",
                        "name": "user_id",
                        "description": "user_id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Sucess"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Authentication failure"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/getUserWithName": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Returns user with matching name ",
                "parameters": [{
                        "in": "header",
                        "name": "Authorization",
                        "description": "Bearer followed by auth token recieved after login",
                        "required": true,
                        "type": "string",
                        "example": "Bearer eyJhbGciOiJIUzI1NiJ9.ZW1haWxAbm9lbWFpbC5jb20.fGbjTahMFHe9Ad330JH5HfkhNuP4FdXICuQHgKS9Xag"
                    },
                    {
                        "in": "query",
                        "name": "user_name",
                        "description": "Book name starting with",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Sucess"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Authentication failure"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Add a new user ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "in": "body",
                        "name": "body",
                        "description": "request body parameters need to be added here",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Authorization",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "example": "Bearer eyJhbGciOiJIUzI1NiJ9.ZW1haWxAbm9lbWFpbC5jb20.fGbjTahMFHe9Ad330JH5HfkhNuP4FdXICuQHgKS9Xag"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucess"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Authentication failure"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update a user ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "in": "body",
                        "name": "body",
                        "description": "request body parameters need to be added here",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateUser"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Authorization",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "example": "Bearer eyJhbGciOiJIUzI1NiJ9.ZW1haWxAbm9lbWFpbC5jb20.fGbjTahMFHe9Ad330JH5HfkhNuP4FdXICuQHgKS9Xag"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucess"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Authentication failure"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete a user ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "in": "query",
                        "name": "user_id",
                        "description": "user_id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Authorization",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "example": "Bearer eyJhbGciOiJIUzI1NiJ9.ZW1haWxAbm9lbWFpbC5jb20.fGbjTahMFHe9Ad330JH5HfkhNuP4FdXICuQHgKS9Xag"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucess"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Authentication failure"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/books": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Returns list of all books",
                "parameters": [{
                    "in": "header",
                    "name": "Authorization",
                    "description": "Bearer followed by auth token recieved after login",
                    "required": true,
                    "type": "string",
                    "example": "Bearer eyJhbGciOiJIUzI1NiJ9.ZW1haWxAbm9lbWFpbC5jb20.fGbjTahMFHe9Ad330JH5HfkhNuP4FdXICuQHgKS9Xag"
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Sucess"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Authentication failure"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/getBookWithId": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Returns book with matching id ",
                "parameters": [{
                        "in": "header",
                        "name": "Authorization",
                        "description": "Bearer followed by auth token recieved after login",
                        "required": true,
                        "type": "string",
                        "example": "Bearer eyJhbGciOiJIUzI1NiJ9.ZW1haWxAbm9lbWFpbC5jb20.fGbjTahMFHe9Ad330JH5HfkhNuP4FdXICuQHgKS9Xag"

                    },
                    {
                        "in": "query",
                        "name": "bookId",
                        "description": "bookId",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Sucess"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Authentication failure"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/getBookWithName": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Returns book with matching name ",
                "parameters": [{
                        "in": "header",
                        "name": "Authorization",
                        "description": "Bearer followed by auth token recieved after login",
                        "required": true,
                        "type": "string",
                        "example": "Bearer eyJhbGciOiJIUzI1NiJ9.ZW1haWxAbm9lbWFpbC5jb20.fGbjTahMFHe9Ad330JH5HfkhNuP4FdXICuQHgKS9Xag"
                    },
                    {
                        "in": "query",
                        "name": "bookName",
                        "description": "Book name starting with",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Sucess"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Authentication failure"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/book": {
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "Add  books ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "in": "body",
                        "name": "body",
                        "description": "request body parameters need to be added here",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Book"
                            }

                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Authorization",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "example": "Bearer eyJhbGciOiJIUzI1NiJ9.ZW1haWxAbm9lbWFpbC5jb20.fGbjTahMFHe9Ad330JH5HfkhNuP4FdXICuQHgKS9Xag"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucess"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Authentication failure"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Book"
                ],
                "summary": "Update a book ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "in": "body",
                        "name": "body",
                        "description": "request body parameters need to be added here",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateBook"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Authorization",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "example": "Bearer eyJhbGciOiJIUzI1NiJ9.ZW1haWxAbm9lbWFpbC5jb20.fGbjTahMFHe9Ad330JH5HfkhNuP4FdXICuQHgKS9Xag"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucess"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Authentication failure"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Book"
                ],
                "summary": "Delete a book ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "in": "query",
                        "name": "bookId",
                        "description": "bookId",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Authorization",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "example": "Bearer eyJhbGciOiJIUzI1NiJ9.ZW1haWxAbm9lbWFpbC5jb20.fGbjTahMFHe9Ad330JH5HfkhNuP4FdXICuQHgKS9Xag"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucess"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Authentication failure"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/publications": {
            "get": {
                "tags": [
                    "Publication"
                ],
                "summary": "Returns list of all publications ",
                "parameters": [{
                    "in": "header",
                    "name": "Authorization",
                    "description": "Bearer followed by auth token recieved after login",
                    "required": true,
                    "type": "string",
                    "example": "Bearer eyJhbGciOiJIUzI1NiJ9.ZW1haWxAbm9lbWFpbC5jb20.fGbjTahMFHe9Ad330JH5HfkhNuP4FdXICuQHgKS9Xag"
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Sucess"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Authentication failure"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "User": {
            "type": "object",
            "required": [
                "name", "email", "password"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": " Name"
                },
                "email": {
                    "type": "string",
                    "example": "email "
                },
                "password": {
                    "type": "string",
                    "example": "password"
                }
            }
        },
        "UpdateUser": {
            "type": "object",
            "required": [
                "name", "user_id"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": " Name"
                },
                "user_id": {
                    "type": "string",
                    "example": "user_id"
                }
            }
        },
        "Login": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "email@noemail.com"
                },
                "password": {
                    "type": "string",
                    "example": "somepassword"
                }
            }
        }
    }
}